Remove kubectl from istioctl
