+++ operator/cmd/mesh/manifest-apply.go
@@ -20,13 +20,10 @@ import (
 	"time"
 
 	"github.com/spf13/cobra"
-	"k8s.io/client-go/kubernetes/scheme"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"istio.io/api/operator/v1alpha1"
 	iopv1alpha1 "istio.io/istio/operator/pkg/apis/istio/v1alpha1"
 	"istio.io/istio/operator/pkg/helmreconciler"
-	"istio.io/istio/operator/pkg/manifest"
 	"istio.io/istio/operator/pkg/object"
 	"istio.io/istio/operator/pkg/translate"
 	"istio.io/istio/operator/pkg/util/clog"
@@ -165,16 +162,7 @@ func ApplyManifests(setOverlay []string, inFilenames []string, force bool, dryRu
 	if err != nil {
 		return err
 	}
-
-	restConfig, clientSet, err := manifest.InitK8SRestClient(kubeConfigPath, context)
-	if err != nil {
-		return err
-	}
-	// We are running a one-off command locally, so we don't need to worry too much about rate limitting
-	// Bumping this up greatly decreases install time
-	restConfig.QPS = 50
-	restConfig.Burst = 100
-	client, err := client.New(restConfig, client.Options{Scheme: scheme.Scheme})
+	restConfig, client, err := K8sConfig(kubeConfigPath, context)
 	if err != nil {
 		return err
 	}
@@ -192,7 +180,7 @@ func ApplyManifests(setOverlay []string, inFilenames []string, force bool, dryRu
 		return err
 	}
 
-	if err := manifest.CreateNamespace(iop.Namespace); err != nil {
+	if err := CreateNamespace(iop.Namespace); err != nil {
 		return err
 	}
 
@@ -209,19 +197,6 @@ func ApplyManifests(setOverlay []string, inFilenames []string, force bool, dryRu
 	}
 	if status.Status != v1alpha1.InstallStatus_HEALTHY {
 		return fmt.Errorf("errors occurred during operation")
-	}
-
-	if wait {
-		l.LogAndPrint("Waiting for resources to become ready...")
-		objs, err := object.ParseK8sObjectsFromYAMLManifest(reconciler.GetManifests().String())
-		if err != nil {
-			l.LogAndPrintf("\n\n✘ Errors in manifest:\n%s\n", err)
-			return fmt.Errorf("errors during wait")
-		}
-		if err := manifest.WaitForResources(objs, clientSet, waitTimeout, dryRun, l); err != nil {
-			l.LogAndPrintf("\n\n✘ Errors during wait:\n%s\n", err)
-			return fmt.Errorf("errors during wait")
-		}
 	}
 
 	l.LogAndPrint("\n\n✔ Installation complete\n")
