+++ operator/cmd/mesh/manifest-migrate.go
@@ -15,7 +15,6 @@
 package mesh
 
 import (
-	"encoding/json"
 	"fmt"
 	"path/filepath"
 
@@ -24,7 +23,6 @@ import (
 	"github.com/spf13/cobra"
 
 	iopv1alpha1 "istio.io/istio/operator/pkg/apis/istio/v1alpha1"
-	"istio.io/istio/operator/pkg/kubectlcmd"
 	"istio.io/istio/operator/pkg/translate"
 	"istio.io/istio/operator/pkg/util"
 	"istio.io/istio/operator/pkg/util/clog"
@@ -55,18 +53,13 @@ func manifestMigrateCmd(rootArgs *rootArgs, mmArgs *manifestMigrateArgs) *cobra.
 		Short: "Migrates a file containing Helm values or IstioControlPlane to IstioOperator format",
 		Long:  "The migrate subcommand migrates a configuration from Helm values or IstioControlPlane format to IstioOperator format.",
 		Args: func(cmd *cobra.Command, args []string) error {
-			if len(args) > 1 {
+			if len(args) != 1 {
 				return fmt.Errorf("migrate accepts optional single filepath")
 			}
 			return nil
 		},
 		RunE: func(cmd *cobra.Command, args []string) error {
 			l := clog.NewConsoleLogger(rootArgs.logToStdErr, cmd.OutOrStdout(), cmd.ErrOrStderr())
-
-			if len(args) == 0 {
-				return migrateFromClusterConfig(rootArgs, mmArgs, l)
-			}
-
 			return migrateFromFiles(rootArgs, mmArgs, args, l)
 		}}
 }
@@ -123,38 +116,4 @@ func translateFunc(values []byte, force bool, l clog.Logger) error {
 
 	l.Print(string(isCPYaml) + "\n")
 	return nil
-}
-
-// migrateFromClusterConfig handles migration for in cluster config.
-func migrateFromClusterConfig(rootArgs *rootArgs, mmArgs *manifestMigrateArgs, l clog.Logger) error {
-	initLogsOrExit(rootArgs)
-
-	l.LogAndPrint("translating in cluster specs\n")
-
-	c := kubectlcmd.New()
-	opts := &kubectlcmd.Options{
-		Namespace: mmArgs.namespace,
-		Output:    "jsonpath='{.data.values}'",
-	}
-	output, stderr, err := c.GetConfigMap("istio-sidecar-injector", opts)
-	if err != nil {
-		return err
-	}
-	if stderr != "" {
-		l.LogAndPrint("error: ", stderr, "\n")
-	}
-	var value map[string]interface{}
-	if len(output) > 1 {
-		output = output[1 : len(output)-1]
-	}
-	err = json.Unmarshal([]byte(output), &value)
-	if err != nil {
-		return fmt.Errorf("error unmarshaling JSON to untyped map %s", err)
-	}
-	res, err := yaml.Marshal(value)
-	if err != nil {
-		return fmt.Errorf("error marshaling untyped map to YAML: %s", err)
-	}
-
-	return translateFunc(res, mmArgs.force, l)
 }
