+++ operator/cmd/mesh/operator-common.go
@@ -15,8 +15,20 @@
 package mesh
 
 import (
+	"context"
+	"fmt"
+	"os"
+
+	"k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/errors"
+	v12 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime/serializer"
+	"k8s.io/client-go/kubernetes"
+	"k8s.io/client-go/kubernetes/scheme"
+	"k8s.io/client-go/rest"
+	"k8s.io/client-go/tools/clientcmd"
+
 	"istio.io/istio/operator/pkg/helm"
-	"istio.io/istio/operator/pkg/manifest"
 	"istio.io/istio/operator/pkg/util"
 	"istio.io/istio/operator/pkg/util/clog"
 )
@@ -39,7 +51,7 @@ const (
 )
 
 func isControllerInstalled(kubeconfig, context, operatorNamespace string) (bool, error) {
-	return manifest.DeploymentExists(kubeconfig, context, operatorNamespace, operatorResourceName)
+	return DeploymentExists(kubeconfig, context, operatorNamespace, operatorResourceName)
 }
 
 // chartsRootDir, helmBaseDir, componentName, namespace string) (Template, TemplateRenderer, error) {
@@ -81,4 +93,114 @@ tag: {{.Tag}}
 	}
 	manifest, err := r.RenderManifest(vals)
 	return vals, manifest, err
+}
+
+var k8sRESTConfig *rest.Config
+
+var k8sClientset *kubernetes.Clientset
+
+var currentKubeconfig string
+
+var currentContext string
+
+func CreateNamespace(namespace string) error {
+	if namespace == "" {
+		// Setup default namespace
+		namespace = "istio-system"
+	}
+
+	// TODO we need to stop creating configs in every function that needs them. One client should be used.
+	cs, e := kubernetes.NewForConfig(k8sRESTConfig)
+	if e != nil {
+		return fmt.Errorf("k8s client error: %s", e)
+	}
+	ns := &v1.Namespace{ObjectMeta: v12.ObjectMeta{
+		Name: namespace,
+		Labels: map[string]string{
+			"istio-injection": "disabled",
+		},
+	}}
+	_, err := cs.CoreV1().Namespaces().Create(context.TODO(), ns, v12.CreateOptions{})
+	if err != nil && !errors.IsAlreadyExists(err) {
+		return fmt.Errorf("failed to create namespace %v: %v", namespace, err)
+	}
+	return nil
+}
+
+func DeploymentExists(kubeconfig, kubeContext, namespace, name string) (bool, error) {
+	if _, _, err := InitK8SRestClient(kubeconfig, kubeContext); err != nil {
+		return false, err
+	}
+
+	cs, err := kubernetes.NewForConfig(k8sRESTConfig)
+	if err != nil {
+		return false, fmt.Errorf("k8s client error: %s", err)
+	}
+
+	d, err := cs.AppsV1().Deployments(namespace).Get(context.TODO(), name, v12.GetOptions{})
+	if err != nil {
+		return false, err
+	}
+	return d != nil, nil
+}
+
+func InitK8SRestClient(kubeconfig, kubeContext string) (*rest.Config, *kubernetes.Clientset, error) {
+	var err error
+	if kubeconfig == currentKubeconfig && kubeContext == currentContext && k8sRESTConfig != nil {
+		return k8sRESTConfig, k8sClientset, nil
+	}
+	currentKubeconfig, currentContext = kubeconfig, kubeContext
+
+	k8sRESTConfig, err = defaultRestConfig(kubeconfig, kubeContext)
+	if err != nil {
+		return nil, nil, err
+	}
+	k8sClientset, err = kubernetes.NewForConfig(k8sRESTConfig)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	return k8sRESTConfig, k8sClientset, nil
+}
+
+func defaultRestConfig(kubeconfig, kubeContext string) (*rest.Config, error) {
+	config, err := BuildClientConfig(kubeconfig, kubeContext)
+	if err != nil {
+		return nil, err
+	}
+	config.APIPath = "/api"
+	config.GroupVersion = &v1.SchemeGroupVersion
+	config.NegotiatedSerializer = serializer.WithoutConversionCodecFactory{CodecFactory: scheme.Codecs}
+	return config, nil
+}
+
+// BuildClientConfig is a helper function that builds client config from a kubeconfig filepath.
+// It overrides the current context with the one provided (empty to use default).
+//
+// This is a modified version of k8s.io/client-go/tools/clientcmd/BuildConfigFromFlags with the
+// difference that it loads default configs if not running in-cluster.
+func BuildClientConfig(kubeconfig, context string) (*rest.Config, error) {
+	if kubeconfig != "" {
+		info, err := os.Stat(kubeconfig)
+		if err != nil || info.Size() == 0 {
+			// If the specified kubeconfig doesn't exists / empty file / any other error
+			// from file stat, fall back to default
+			kubeconfig = ""
+		}
+	}
+
+	//Config loading rules:
+	// 1. kubeconfig if it not empty string
+	// 2. In cluster config if running in-cluster
+	// 3. Config(s) in KUBECONFIG environment variable
+	// 4. Use $HOME/.kube/config
+	loadingRules := clientcmd.NewDefaultClientConfigLoadingRules()
+	loadingRules.DefaultClientConfig = &clientcmd.DefaultClientConfig
+	loadingRules.ExplicitPath = kubeconfig
+	configOverrides := &clientcmd.ConfigOverrides{
+		ClusterDefaults: clientcmd.ClusterDefaults,
+		CurrentContext:  context,
+	}
+
+	return clientcmd.NewNonInteractiveDeferredLoadingClientConfig(loadingRules, configOverrides).ClientConfig()
 }
