+++ operator/cmd/mesh/upgrade.go
@@ -15,18 +15,21 @@
 package mesh
 
 import (
+	"context"
 	"fmt"
 	"os"
 	"time"
 
+	"github.com/docker/distribution/reference"
 	goversion "github.com/hashicorp/go-version"
 	"github.com/spf13/cobra"
+	"k8s.io/api/core/v1"
+	"k8s.io/client-go/rest"
 
 	iop "istio.io/istio/operator/pkg/apis/istio/v1alpha1"
 	"istio.io/istio/operator/pkg/compare"
-	"istio.io/istio/operator/pkg/hooks"
-	"istio.io/istio/operator/pkg/manifest"
 	"istio.io/istio/operator/pkg/tpath"
+	"istio.io/istio/operator/pkg/util"
 	"istio.io/istio/operator/pkg/util/clog"
 	pkgversion "istio.io/istio/operator/pkg/version"
 	"istio.io/pkg/log"
@@ -124,7 +127,7 @@ func UpgradeCmd() *cobra.Command {
 // upgrade is the main function for Upgrade command
 func upgrade(rootArgs *rootArgs, args *upgradeArgs, l clog.Logger) (err error) {
 	// Create a kube client from args.kubeConfigPath and  args.context
-	kubeClient, err := manifest.NewClient(args.kubeConfigPath, args.context)
+	kubeClient, err := NewClient(args.kubeConfigPath, args.context)
 	if err != nil {
 		return fmt.Errorf("failed to connect Kubernetes API server, error: %v", err)
 	}
@@ -202,29 +205,11 @@ func upgrade(rootArgs *rootArgs, args *upgradeArgs, l clog.Logger) (err error) {
 
 	waitForConfirmation(args.skipConfirmation, l)
 
-	// Run pre-upgrade hooks
-	hparams := &hooks.HookCommonParams{
-		SourceVer:  currentVersion,
-		TargetVer:  targetVersion,
-		SourceIOPS: targetIOPS,
-		TargetIOPS: targetIOPS,
-	}
-	errs := hooks.RunPreUpgradeHooks(kubeClient, hparams, rootArgs.dryRun)
-	if len(errs) != 0 && !args.force {
-		return fmt.Errorf("failed in pre-upgrade hooks, error: %v", errs.ToError())
-	}
-
 	// Apply the Istio Control Plane specs reading from inFilenames to the cluster
 	err = ApplyManifests(nil, args.inFilenames, args.force, rootArgs.dryRun,
 		rootArgs.verbose, args.kubeConfigPath, args.context, args.wait, upgradeWaitSecWhenApply, l)
 	if err != nil {
 		return fmt.Errorf("failed to apply the Istio Control Plane specs. Error: %v", err)
-	}
-
-	// Run post-upgrade hooks
-	errs = hooks.RunPostUpgradeHooks(kubeClient, hparams, rootArgs.dryRun)
-	if len(errs) != 0 && !args.force {
-		return fmt.Errorf("failed in post-upgrade hooks, error: %v", errs.ToError())
 	}
 
 	if !args.wait {
@@ -302,7 +287,7 @@ func checkSupportedVersions(cur, tar, versionsURI string) error {
 }
 
 // retrieveControlPlaneVersion retrieves the version number from the Istio control plane
-func retrieveControlPlaneVersion(kubeClient manifest.ExecClient, istioNamespace string, l clog.Logger) (string, error) {
+func retrieveControlPlaneVersion(kubeClient ExecClient, istioNamespace string, l clog.Logger) (string, error) {
 	cv, e := kubeClient.GetIstioVersions(istioNamespace)
 	if e != nil {
 		return "", fmt.Errorf("failed to retrieve Istio control plane version, error: %v", e)
@@ -326,7 +311,7 @@ func retrieveControlPlaneVersion(kubeClient manifest.ExecClient, istioNamespace 
 
 // waitUpgradeComplete waits for the upgrade to complete by periodically comparing the current component version
 // to the target version.
-func waitUpgradeComplete(kubeClient manifest.ExecClient, istioNamespace string, targetVer string, l clog.Logger) error {
+func waitUpgradeComplete(kubeClient ExecClient, istioNamespace string, targetVer string, l clog.Logger) error {
 	for i := 1; i <= upgradeWaitCheckVerMaxAttempts; i++ {
 		sleepSeconds(upgradeWaitSecCheckVerPerLoop)
 		cv, e := kubeClient.GetIstioVersions(istioNamespace)
@@ -363,7 +348,7 @@ func sleepSeconds(duration time.Duration) {
 }
 
 // coalesceVersions coalesces all Istio control plane components versions
-func coalesceVersions(cv []manifest.ComponentVersion) (string, error) {
+func coalesceVersions(cv []ComponentVersion) (string, error) {
 	if len(cv) == 0 {
 		return "", fmt.Errorf("empty list of ComponentVersion")
 	}
@@ -374,7 +359,7 @@ func coalesceVersions(cv []manifest.ComponentVersion) (string, error) {
 }
 
 // identicalVersions checks if Istio control plane components are on the same version
-func identicalVersions(cv []manifest.ComponentVersion) bool {
+func identicalVersions(cv []ComponentVersion) bool {
 	exemplar := cv[0]
 	for i := 1; i < len(cv); i++ {
 		if exemplar.Version != cv[i].Version {
@@ -382,4 +367,150 @@ func identicalVersions(cv []manifest.ComponentVersion) bool {
 		}
 	}
 	return true
+}
+
+// Client is a helper wrapper around the Kube RESTClient for istioctl -> Pilot/Envoy/Mesh related things
+type Client struct {
+	Config *rest.Config
+	*rest.RESTClient
+}
+
+// ComponentVersion is a pair of component name and version
+type ComponentVersion struct {
+	Component string
+	Version   string
+	Pod       v1.Pod
+}
+
+func (cv ComponentVersion) String() string {
+	return fmt.Sprintf("%s pod - %s - version: %s",
+		cv.Component, cv.Pod.GetName(), cv.Version)
+}
+
+// ExecClient is an interface for remote execution
+type ExecClient interface {
+	GetIstioVersions(namespace string) ([]ComponentVersion, error)
+	GetPods(namespace string, params map[string]string) (*v1.PodList, error)
+	PodsForSelector(namespace, labelSelector string) (*v1.PodList, error)
+	ConfigMapForSelector(namespace, labelSelector string) (*v1.ConfigMapList, error)
+}
+
+// NewClient is the constructor for the client wrapper
+func NewClient(kubeconfig, configContext string) (*Client, error) {
+	config, err := defaultRestConfig(kubeconfig, configContext)
+	if err != nil {
+		return nil, err
+	}
+	restClient, err := rest.RESTClientFor(config)
+	if err != nil {
+		return nil, err
+	}
+	return &Client{config, restClient}, nil
+}
+
+// GetIstioVersions gets the version for each Istio component
+func (client *Client) GetIstioVersions(namespace string) ([]ComponentVersion, error) {
+	pods, err := client.GetPods(namespace, map[string]string{
+		"labelSelector": "istio",
+		"fieldSelector": "status.phase=Running",
+	})
+	if err != nil {
+		return nil, fmt.Errorf("failed to retrieve Istio pods, error: %v", err)
+	}
+	if len(pods.Items) == 0 {
+		return nil, fmt.Errorf("istio pod not found in namespace %v", namespace)
+	}
+
+	var errs util.Errors
+	var res []ComponentVersion
+	for _, pod := range pods.Items {
+		component := pod.Labels["istio"]
+
+		switch component {
+		case "statsd-prom-bridge":
+			continue
+		case "mixer":
+			component = pod.Labels["istio-mixer-type"]
+		}
+
+		server := ComponentVersion{
+			Component: component,
+			Pod:       pod,
+		}
+
+		pv := ""
+		for _, c := range pod.Spec.Containers {
+			cv, err := parseTag(c.Image)
+			if err != nil {
+				errs = util.AppendErr(errs, err)
+			}
+
+			if pv == "" {
+				pv = cv
+			} else if pv != cv {
+				err := fmt.Errorf("differrent versions of containers in the same pod: %v", pod.Spec.Containers)
+				errs = util.AppendErr(errs, err)
+			}
+		}
+		server.Version, err = pkgversion.TagToVersionString(pv)
+		if err != nil {
+			tagErr := fmt.Errorf("unable to convert tag %s into version in pod: %v", pv, pod.Spec.Containers)
+			errs = util.AppendErr(errs, tagErr)
+		}
+		res = append(res, server)
+	}
+	return res, errs.ToError()
+}
+
+func parseTag(image string) (string, error) {
+	ref, err := reference.Parse(image)
+	if err != nil {
+		return "", fmt.Errorf("could not parse image: %s, error: %v", image, err)
+	}
+
+	switch t := ref.(type) {
+	case reference.Tagged:
+		return t.Tag(), nil
+	default:
+		return "", fmt.Errorf("tag not found in image: %v", image)
+	}
+}
+
+func (client *Client) PodsForSelector(namespace, labelSelector string) (*v1.PodList, error) {
+	pods, err := client.GetPods(namespace, map[string]string{
+		"labelSelector": labelSelector,
+	})
+	if err != nil {
+		return nil, fmt.Errorf("failed to retrieve pods, error: %v", err)
+	}
+	return pods, nil
+}
+
+// GetPods retrieves the pod objects for Istio deployments
+func (client *Client) GetPods(namespace string, params map[string]string) (*v1.PodList, error) {
+	req := client.Get().
+		Resource("pods").
+		Namespace(namespace)
+	for k, v := range params {
+		req.Param(k, v)
+	}
+
+	res := req.Do(context.TODO())
+	if res.Error() != nil {
+		return nil, fmt.Errorf("unable to retrieve Pods: %v", res.Error())
+	}
+	list := &v1.PodList{}
+	if err := res.Into(list); err != nil {
+		return nil, fmt.Errorf("unable to parse PodList: %v", res.Error())
+	}
+	return list, nil
+}
+
+func (client *Client) ConfigMapForSelector(namespace, labelSelector string) (*v1.ConfigMapList, error) {
+	cmGet := client.Get().Resource("configmaps").Namespace(namespace).Param("labelSelector", labelSelector)
+	obj, err := cmGet.Do(context.TODO()).Get()
+	if err != nil {
+		return nil, fmt.Errorf("failed retrieving configmap: %v", err)
+	}
+	return obj.(*v1.ConfigMapList), nil
 }
