+++ operator/pkg/helmreconciler/prune.go
@@ -20,11 +20,10 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime/schema"
-	utilerrors "k8s.io/apimachinery/pkg/util/errors"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
-	"istio.io/istio/operator/pkg/apis/istio/v1alpha1"
 	"istio.io/istio/operator/pkg/object"
+	"istio.io/istio/operator/pkg/util"
 )
 
 const (
@@ -33,10 +32,6 @@ const (
 
 	// OwnerNameKey represents the name of the owner to which the resource relates
 	OwnerNameKey = MetadataNamespace + "/owner-name"
-	// OwnerKindKey represents the kind of the owner to which the resource relates
-	OwnerKindKey = MetadataNamespace + "/owner-kind"
-	// OwnerGroupKey represents the group of the owner to which the resource relates
-	OwnerGroupKey = MetadataNamespace + "/owner-group"
 )
 
 var (
@@ -98,47 +93,32 @@ var (
 	}
 )
 
-// NewPruningDetails creates a new PruningDetails object specific to the instance.
-func NewIstioPruningDetails(instance *v1alpha1.IstioOperator) PruningDetails {
-	name := instance.GetName()
-	return &SimplePruningDetails{
-		OwnerLabels: map[string]string{
-			OwnerNameKey:  name,
-			OwnerGroupKey: v1alpha1.IstioOperatorGVK.Group,
-			OwnerKindKey:  v1alpha1.IstioOperatorGVK.Kind,
-		},
-		NamespacedResources:    namespacedResources,
-		NonNamespacedResources: nonNamespacedResources,
+// Prune removes any resources not specified in manifests generated by HelmReconciler h.
+func (h *HelmReconciler) Prune(manifests ChartManifestsMap) error {
+	var errs util.Errors
+	for cname, manifest := range manifests.Consolidated() {
+		errs = util.AppendErr(errs, h.PruneUnlistedResources(allObjectHashes(manifest), cname))
 	}
+	return errs.ToError()
 }
 
-// Prune removes any resources not specified in manifests generated by HelmReconciler h. If all is set to true, this
-// function prunes all resources.
-func (h *HelmReconciler) Prune(excluded map[string]bool, all bool) error {
-	namespacedResources, clusterResources := h.pruningDetails.GetResourceTypes()
-	targetNamespace := h.iop.Namespace
-	err := h.PruneUnlistedResources(append(namespacedResources, clusterResources...), excluded, all, targetNamespace)
-	if all {
-		FlushObjectCaches()
-	}
-	return err
-}
-
-func (h *HelmReconciler) PruneUnlistedResources(gvks []schema.GroupVersionKind, excluded map[string]bool, all bool, namespace string) error {
-	allErrors := []error{}
-	ownerLabels := h.pruningDetails.GetOwnerLabels()
-	for _, gvk := range gvks {
+func (h *HelmReconciler) PruneUnlistedResources(excluded map[string]bool, componentName string) error {
+	var errs util.Errors
+	for _, gvk := range append(namespacedResources, nonNamespacedResources...) {
 		objects := &unstructured.UnstructuredList{}
 		objects.SetGroupVersionKind(gvk)
-		err := h.client.List(context.TODO(), objects, client.MatchingLabels(ownerLabels), client.InNamespace(namespace))
+		labels, err := h.getOwnerLabels(componentName)
 		if err != nil {
+			return err
+		}
+		if err := h.client.List(context.TODO(), objects, client.MatchingLabels(labels), client.InNamespace(h.iop.Namespace)); err != nil {
 			// we only want to retrieve resources clusters
 			scope.Warnf("retrieving resources to prune type %s: %s not found", gvk.String(), err)
 			continue
 		}
 		for _, o := range objects.Items {
 			oh := object.NewK8sObject(&o, nil, nil).Hash()
-			if excluded[oh] && !all {
+			if excluded[oh] {
 				continue
 			}
 			if h.opts.DryRun {
@@ -148,11 +128,10 @@ func (h *HelmReconciler) PruneUnlistedResources(gvks []schema.GroupVersionKind, 
 
 			err = h.client.Delete(context.TODO(), &o, client.PropagationPolicy(metav1.DeletePropagationBackground))
 			if err != nil {
-				allErrors = append(allErrors, err)
+				errs = util.AppendErr(errs, err)
 			}
 			h.opts.Log.LogAndPrintf("Pruned object %s.", oh)
-
 		}
 	}
-	return utilerrors.NewAggregate(allErrors)
+	return errs.ToError()
 }
