+++ operator/pkg/helmreconciler/reconcile.go
@@ -28,7 +28,6 @@ import (
 
 	"istio.io/api/operator/v1alpha1"
 	valuesv1alpha1 "istio.io/istio/operator/pkg/apis/istio/v1alpha1"
-	"istio.io/istio/operator/pkg/manifest"
 	"istio.io/istio/operator/pkg/name"
 	"istio.io/istio/operator/pkg/object"
 	"istio.io/istio/operator/pkg/util/clog"
@@ -73,13 +72,11 @@ func init() {
 
 // HelmReconciler reconciles resources rendered by a set of helm charts.
 type HelmReconciler struct {
-	client             client.Client
-	restConfig         *rest.Config
-	clientSet          *kubernetes.Clientset
-	iop                *valuesv1alpha1.IstioOperator
-	pruningDetails     PruningDetails
-	opts               *Options
-	needUpdateAndPrune bool
+	client     client.Client
+	restConfig *rest.Config
+	clientSet  *kubernetes.Clientset
+	iop        *valuesv1alpha1.IstioOperator
+	opts       *Options
 	// copy of the last generated manifests.
 	manifests name.ManifestMap
 }
@@ -90,9 +87,14 @@ type Options struct {
 	DryRun bool
 	// Log is a console logger for user visible CLI output.
 	Log clog.Logger
+	// WaitTimeout is the maximum amount of time to wait for resources to become ready.
+	WaitTimeout time.Duration
 }
 
-var defaultOptions = &Options{Log: clog.NewDefaultLogger()}
+var defaultOptions = &Options{
+	Log:         clog.NewDefaultLogger(),
+	WaitTimeout: internalDepTimeout,
+}
 
 // NewHelmReconciler creates a HelmReconciler and returns a ptr to it
 func NewHelmReconciler(client client.Client, restConfig *rest.Config, iop *valuesv1alpha1.IstioOperator, opts *Options) (*HelmReconciler, error) {
@@ -108,13 +110,11 @@ func NewHelmReconciler(client client.Client, restConfig *rest.Config, iop *value
 		return nil, err
 	}
 	return &HelmReconciler{
-		client:             client,
-		restConfig:         restConfig,
-		clientSet:          cs,
-		iop:                iop,
-		pruningDetails:     NewIstioPruningDetails(iop),
-		opts:               opts,
-		needUpdateAndPrune: true,
+		client:     client,
+		restConfig: restConfig,
+		clientSet:  cs,
+		iop:        iop,
+		opts:       opts,
 	}, nil
 }
 
@@ -125,14 +125,7 @@ func (h *HelmReconciler) Reconcile() (*v1alpha1.InstallStatus, error) {
 		return nil, err
 	}
 
-	status := h.processRecursive(manifestMap)
-
-	// Delete any resources not in the manifest but managed by operator.
-	if h.needUpdateAndPrune {
-		err = h.Prune(allObjectHashes(manifestMap), false)
-	}
-
-	return status, err
+	return h.processRecursive(manifestMap), h.Prune(manifestMap)
 }
 
 // processRecursive processes the given manifests in an order of dependencies defined in h. Dependencies are a tree,
@@ -178,14 +171,6 @@ func (h *HelmReconciler) processRecursive(manifests ChartManifestsMap) *v1alpha1
 			setStatus(componentStatus, c, status, err)
 			mu.Unlock()
 
-			// If we are depending on a component, we may depend on it actually running (eg Deployment is ready)
-			// For example, for the validation webhook to become ready, so we should wait for it always.
-			if err == nil && len(componentDependencies[cn]) > 0 {
-				if err := manifest.WaitForResources(processedObjs, h.clientSet, internalDepTimeout, h.opts.DryRun, h.opts.Log); err != nil {
-					scope.Errorf("Failed to wait for resource: %v", err)
-				}
-			}
-
 			// Signal all the components that depend on us.
 			for _, ch := range componentDependencies[cn] {
 				scope.Infof("Unblocking dependency %s.", ch)
@@ -205,8 +190,11 @@ func (h *HelmReconciler) processRecursive(manifests ChartManifestsMap) *v1alpha1
 
 // Delete resources associated with the custom resource instance
 func (h *HelmReconciler) Delete() error {
-	h.needUpdateAndPrune = true
-	return h.Prune(nil, true)
+	manifestMap, err := h.RenderCharts()
+	if err != nil {
+		return err
+	}
+	return h.Prune(manifestMap)
 }
 
 // SetStatusBegin updates the status field on the IstioOperator instance before reconciling.
@@ -292,19 +280,16 @@ func overallStatus(componentStatus map[string]*v1alpha1.InstallStatus_VersionSta
 }
 
 // allObjectHashes returns a map with object hashes of all the objects contained in cmm as the keys.
-func allObjectHashes(cmm ChartManifestsMap) map[string]bool {
+func allObjectHashes(m string) map[string]bool {
 	ret := make(map[string]bool)
-	for _, mm := range cmm {
-		for _, m := range mm {
-			objs, err := object.ParseK8sObjectsFromYAMLManifest(m.Content)
-			if err != nil {
-				scope.Error(err.Error())
-			}
-			for _, o := range objs {
-				ret[o.Hash()] = true
-			}
-		}
+	objs, err := object.ParseK8sObjectsFromYAMLManifest(m)
+	if err != nil {
+		scope.Error(err.Error())
 	}
+	for _, o := range objs {
+		ret[o.Hash()] = true
+	}
+
 	return ret
 }
